Client fetchinfo command now works.

+++ ITVC.CSharp.Client/FetchInfoDetails.cs
@@ -9,10 +9,10 @@ namespace ITVC.CSharp.Client
     {
         public string VideoCodec { get; set; }
         public string AudioCodec { get; set; }
-        public int FPS { get; set; }
+        public float FPS { get; set; }
         public int Width { get; set; }
         public int Height { get; set; }
-        public int Length { get; set; }
+        public float Length { get; set; }
         public int FileSize { get; set; }
         public int VideoBitRate { get; set; }
